+++ k_lambda_spec/primitives/generic_functions_spec.rb
@@ -1,0 +1,7 @@
+require 'spec_helper'
+
+describe 'Primitives for Generic Functions' do
+  describe 'freeze' do
+    # Should the frozen result be re-evaluated on each thaw?
+  end
+end